refact: changes made after the review
