--- src/apps/core/services/spreadsheets/exceptions.py
@@ -1,6 +1,0 @@
-class SendFailed(Exception):
-    """Не удалось отправить строку в таблицу."""
-
-
-class APIConnectFailed(Exception):
-    """Не удалось подключиться к Google API."""
